move functions from __main__ to __init__
