+++ setup.py
@@ -93,7 +93,7 @@ setup(name='psy-view',
       tests_require=['pytest', 'pytest-qt'],
       cmdclass=cmdclass,
       entry_points={
-          'console_scripts': ['psy-view=psy_view.__main__:main'],
+          'console_scripts': ['psy-view=psy_view:main'],
           'psyplot_gui': ['psy-view=psy_view.ds_widget:DatasetWidgetPlugin'],
           },
       zip_safe=False)
